[ntuple] make RClusterSize_t 64bit

+++ tree/ntuple/v7/src/RPageStorage.cxx
@@ -497,7 +497,7 @@ ROOT::Experimental::Detail::RPageSink::SealPage(const RPage &page,
 
    R__ASSERT(isAdoptedBuffer);
 
-   return RSealedPage{pageBuf, zippedBytes, page.GetNElements()};
+   return RSealedPage{pageBuf, static_cast<std::uint32_t>(zippedBytes), page.GetNElements()};
 }
 
 ROOT::Experimental::Detail::RPageStorage::RSealedPage
